backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
holidays==0.34
pydantic==2.5.0
python-dotenv==1.0.0
redis==4.6.0
websockets>=9.0,<11
fastapi-cache2[redis]==0.2.1
httpx==0.25.2
requests==2.31.0
backoff==2.2.1
ccxt==4.1.77
websocket-client==1.6.4
aiofiles==23.2.1
python-socketio==5.10.0
python-engineio==4.8.0
apscheduler==3.10.4
pandas==2.1.4
numpy==1.25.2
scipy==1.11.3
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
structlog==23.2.0
alpaca-trade-api==3.0.2
aiomysql==0.2.0
greenlet==3.0.3
psycopg2-binary==2.9.9
